LayoutTests/webgpu/whlsl-read-modify-write-expected.txt

PASS results[0] is 2
PASS results[1] is 1
PASS results[2] is 2
PASS results[3] is 2
PASS results[4] is 6
PASS results[5] is 7
PASS results[6] is 34
PASS results[7] is 59
PASS successfullyParsed is true

TEST COMPLETE
